Refactor database package and start auth
