--- database.json
@@ -1,1 +1,0 @@
-{"chirps":{"1":{"body":"I had something interesting for breakfast","id":1},"2":{"body":"What about second breakfast?","id":2}},"users":null}
