+++ go.mod
@@ -3,3 +3,5 @@ module github.com/ajpotts01/go-chirpy
 go 1.21.0
 
 require github.com/go-chi/chi/v5 v5.0.10
+
+require golang.org/x/crypto v0.12.0 // indirect
